Move ABC suggestions to the core versions so they're implemented by default
