+++ src/Parser/Hunter/Marksmanship/Modules/Features/AlwaysBeCasting.js
@@ -7,6 +7,7 @@ import SPELLS from 'common/SPELLS';
 import { formatPercentage } from 'common/format';
 import { STATISTIC_ORDER } from 'Main/StatisticBox';
 import SpellLink from 'common/SpellLink';
+import Wrapper from 'common/Wrapper';
 
 class AlwaysBeCasting extends CoreAlwaysBeCasting {
   static dependencies = {
@@ -15,7 +16,6 @@ class AlwaysBeCasting extends CoreAlwaysBeCasting {
   };
 
   static ABILITIES_ON_GCD = [
-
     // Marksmanship:
 
     // Rotational
@@ -52,14 +52,10 @@ class AlwaysBeCasting extends CoreAlwaysBeCasting {
     SPELLS.CAMOUFLAGE_TALENT.id,
   ];
 
-  get downtime() {
-    return this.totalTimeWasted / this.owner.fightDuration;
-  }
-
   get suggestionThresholds() {
     if (this.combatants.selected.hasTalent(SPELLS.SIDEWINDERS_TALENT.id)) {
       return {
-        actual: this.downtime,
+        actual: this.downtimePercentage,
         isGreaterThan: {
           minor: 0.3,
           average: 0.35,
@@ -69,7 +65,7 @@ class AlwaysBeCasting extends CoreAlwaysBeCasting {
       };
     } else {
       return {
-        actual: this.downtime,
+        actual: this.downtimePercentage,
         isGreaterThan: {
           minor: 0.02,
           average: 0.04,
@@ -79,9 +75,11 @@ class AlwaysBeCasting extends CoreAlwaysBeCasting {
       };
     }
   }
+  statisticOrder = STATISTIC_ORDER.CORE(1);
 
   suggestions(when) {
     const {
+      actual,
       isGreaterThan: {
         minor,
         average,
@@ -90,18 +88,18 @@ class AlwaysBeCasting extends CoreAlwaysBeCasting {
     } = this.suggestionThresholds;
     //When playing with sidewinders your downtime is significantly different than when you play without, this is due to losing all instant casts
     if (this.combatants.selected.hasTalent(SPELLS.SIDEWINDERS_TALENT.id)) {
-      when(this.downtime).isGreaterThan(minor)
+      when(actual).isGreaterThan(minor)
         .addSuggestion((suggest, actual, recommended) => {
-          return suggest(<span>Your downtime can be improved. Try and minimise the time spent standing around doing nothing, even though you are playing with <SpellLink id={SPELLS.SIDEWINDERS_TALENT.id} />.</span>)
+          return suggest(<Wrapper>Your downtime can be improved. Try and minimise the time spent standing around doing nothing, even though you are playing with <SpellLink id={SPELLS.SIDEWINDERS_TALENT.id} />.</Wrapper>)
             .icon('spell_mage_altertime')
             .actual(`${formatPercentage(actual)}% downtime`)
             .recommended(`<${formatPercentage(recommended)}% is recommended`)
             .regular(average).major(major);
         });
     } else {
-      when(this.downtime).isGreaterThan(minor)
+      when(actual).isGreaterThan(minor)
         .addSuggestion((suggest, actual, recommended) => {
-          return suggest(<span>Your downtime can be improved. Try to Always Be Casting (ABC), this means you should try to reduce the delay between casting spells. Even if you have to move, try casting something instant like <SpellLink id={SPELLS.ARCANE_SHOT.id} /> for single target or <SpellLink id={SPELLS.MULTISHOT.id} /> for multiple targets.</span>)
+          return suggest(<Wrapper>Your downtime can be improved. Try to Always Be Casting (ABC), this means you should try to reduce the delay between casting spells. Even if you have to move, try casting something instant like <SpellLink id={SPELLS.ARCANE_SHOT.id} /> for single target or <SpellLink id={SPELLS.MULTISHOT.id} /> for multiple targets.</Wrapper>)
             .icon('spell_mage_altertime')
             .actual(`${formatPercentage(actual)}% downtime`)
             .recommended(`<${formatPercentage(recommended)}% is recommended`)
@@ -109,8 +107,6 @@ class AlwaysBeCasting extends CoreAlwaysBeCasting {
         });
     }
   }
-
-  statisticOrder = STATISTIC_ORDER.CORE(1);
 }
 
 export default AlwaysBeCasting;
